graphisme

+++ src/Vue/Image.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
  */
 public class Image extends JPanel {
     private ArrayList<ArrayList<Integer>> pixelData;
-    private int scaleFactor;
+    private float scaleFactor;
     private float palette;
     private Color color;
 
@@ -22,12 +22,12 @@ public class Image extends JPanel {
      * @param palette     Palette de couleurs
      * @param color       Couleur
      */
-    public Image(ArrayList<ArrayList<Integer>> pixelData, int scaleFactor, float palette, Color color) {
+    public Image(ArrayList<ArrayList<Integer>> pixelData, float scaleFactor, float palette, Color color) {
         this.pixelData = pixelData;
         this.scaleFactor = scaleFactor;
         this.palette = palette;
-        int width = pixelData.get(0).size() * scaleFactor;
-        int height = pixelData.size() * scaleFactor;
+        int width = (int) (pixelData.get(0).size() * scaleFactor);
+        int height = (int) (pixelData.size() * scaleFactor);
         this.color = color;
         setPreferredSize(new Dimension(width, height));
     }
@@ -45,12 +45,12 @@ public class Image extends JPanel {
                 int pixel = pixelData.get(i).get(j);
                 if (pixel == 0) {
                     g.setColor(color);
-                    g.fillRect(j * scaleFactor, i * scaleFactor, scaleFactor, scaleFactor);
+                    g.fillRect((int) (j * scaleFactor), (int)(i * scaleFactor), (int)scaleFactor, (int)scaleFactor);
                     continue;
                 }
                 int gray = (int) (255.0 * ((palette - pixel) / palette));
                 g.setColor(new Color(gray, gray, gray));
-                g.fillRect(j * scaleFactor, i * scaleFactor, scaleFactor, scaleFactor);
+                g.fillRect((int) (j * scaleFactor), (int)(i * scaleFactor), (int)scaleFactor, (int)scaleFactor);
             }
         }
     }
